--- a/options/options.c
+++ b/options/options.c
@@ -90,6 +90,7 @@ extern const struct m_obj_list ao_obj_list;
 
 extern const struct m_sub_options opengl_conf;
 extern const struct m_sub_options vulkan_conf;
+extern const struct m_sub_options spirv_conf;
 extern const struct m_sub_options angle_conf;
 extern const struct m_sub_options cocoa_conf;
 
@@ -686,6 +687,7 @@ const m_option_t mp_opts[] = {
     OPT_SUBSTRUCT("", demux_opts, demux_conf, 0),
 
     OPT_SUBSTRUCT("", gl_video_opts, gl_video_conf, 0),
+    OPT_SUBSTRUCT("", spirv_opts, spirv_conf, 0),
 
 #if HAVE_GL
     OPT_SUBSTRUCT("", opengl_opts, opengl_conf, 0),