--- a/options/options.c
+++ b/options/options.c
@@ -89,6 +89,7 @@ extern const struct m_obj_list vo_obj_list;
 extern const struct m_obj_list ao_obj_list;
 
 extern const struct m_sub_options opengl_conf;
+extern const struct m_sub_options vulkan_conf;
 extern const struct m_sub_options angle_conf;
 extern const struct m_sub_options cocoa_conf;
 
@@ -688,6 +689,10 @@ const m_option_t mp_opts[] = {
 
 #if HAVE_GL
     OPT_SUBSTRUCT("", opengl_opts, opengl_conf, 0),
+#endif
+
+#if HAVE_VULKAN
+    OPT_SUBSTRUCT("", vulkan_opts, vulkan_conf, 0),
 #endif
 
 #if HAVE_EGL_ANGLE_WIN32