--- a/options/options.h
+++ b/options/options.h
@@ -329,6 +329,7 @@ typedef struct MPOpts {
     struct gl_video_opts *gl_video_opts;
     struct angle_opts *angle_opts;
     struct opengl_opts *opengl_opts;
+    struct vulkan_opts *vulkan_opts;
     struct cocoa_opts *cocoa_opts;
     struct dvd_opts *dvd_opts;