--- a/video/out/gpu/context.c
+++ b/video/out/gpu/context.c
@@ -31,6 +31,7 @@
 #include "video/out/vo.h"
 
 #include "context.h"
+#include "spirv.h"
 
 /* OpenGL */
 extern const struct ra_ctx_fns ra_ctx_glx;
@@ -194,10 +195,17 @@ struct ra_ctx *ra_ctx_create(struct vo *vo, const char *context_type,
     return NULL;
 }
 
-void ra_ctx_destroy(struct ra_ctx **ctx)
+void ra_ctx_destroy(struct ra_ctx **ctx_ptr)
 {
-    if (*ctx)
-        (*ctx)->fns->uninit(*ctx);
-    talloc_free(*ctx);
-    *ctx = NULL;
+    struct ra_ctx *ctx = *ctx_ptr;
+    if (!ctx)
+        return;
+
+    if (ctx->spirv && ctx->spirv->fns->uninit)
+        ctx->spirv->fns->uninit(ctx);
+
+    ctx->fns->uninit(ctx);
+    talloc_free(ctx);
+
+    *ctx_ptr = NULL;
 }