--- a/video/out/vulkan/common.h
+++ b/video/out/vulkan/common.h
@@ -49,6 +49,7 @@ struct mpvk_ctx {
     struct vk_cmdpool *pool; // primary command pool for this device
     struct vk_cmdpool *pool_transfer; // pool for async transfers (optional)
     struct vk_cmd *last_cmd; // most recently submitted command
+    struct spirv_compiler *spirv; // GLSL -> SPIR-V compiler
 
     // Cached capabilities
     VkPhysicalDeviceLimits limits;