… DOCS/man/options.rst
… video/out/gpu/spirv_shaderc.c
… video/out/vo_gpu.c
… video/out/vulkan/context.c
… video/out/vulkan/context.h
… video/out/vulkan/context_wayland.c
… video/out/vulkan/context_xlib.c
… video/out/vulkan/formats.c
… video/out/vulkan/formats.h
… video/out/vulkan/malloc.c
… video/out/vulkan/malloc.h
… video/out/vulkan/ra_vk.c
… video/out/vulkan/ra_vk.h
… video/out/vulkan/spirv_nvidia.c
… video/out/vulkan/utils.c
… video/out/vulkan/utils.h
… wscript
… wscript_build.py