We might have to go through some more iterations, but in this form, vulkan is actually usable and stable enough for master. I have a lot of experimental work in other branches though that I don't feel are mature enough for this PR. This doesn't represent the limit of what's possible with vulkan by any means, but it should hopefully at least get us some basic testing out there. Essentially, the current version of vulkan is an "opengl simulator". We don't really use any of the super-advanced vulkan features yet, which means the performance is, at best, about "as good" as OpengGL. But that's good enough for starters, I think.

There's at least one issue that needs to be fixed still: The memory allocator makes some assumptions that aren't true for mesa/intel. Easily changed, but needs to be done. There'll probably be more breakage on intel that we have to fix as well. (Also, I think there are some potential use-after-free bugs in the lazy object destruction system, which I'll also get around to fixing soon-ish)

Anyway, it's probably time for some initial code review, since the initial version of vulkan (pending the potential mega-refactors) is probably going to remain looking like this. There will be a lot of code to go through either way.