@haasn 2017-09-22
cc @rossy windows support shouldn't be that difficult to add