@haasn 2017-09-23
> The memory allocator makes some assumptions that aren't true for mesa/intel. Easily changed, but needs to be done. There'll probably be more breakage on intel that we have to fix as well. (Also, I think there are some potential use-after-free bugs in the lazy object destruction system, which I'll also get around to fixing soon-ish)

Fixed both