@haasn 2017-09-23
> Essentially, the current version of vulkan is an "opengl simulator". We don't really use any of the super-advanced vulkan features yet, which means the performance is, at best, about "as good" as OpengGL. But that's good enough for starters, I think.

I just pushed a commit that allows us to properly track image dependencies across frames on a per-renderpass granularity. (This required splitting up our monolithic command buffers into per-pass command buffers [1])

This allows me to safely enable async transfer, boosting performance anywhere from 10% to 25%, and making vulkan quite definitely faster than OpenGL for me: on 1080p clips I get like 410 -> 540 fps for opengl -> vulkan, and on the 8K torture test I get 38 -> 45 fps. (Compared to 35 fps with vulkan before this commit)

[1] Incidentally, this change  we could probably try and avoid re-recording command buffers per frame if we wanted to, potentially gaining some minor CPU reductions. Not a huge deal, though.