@haasn 2017-09-24
I pushed another batch of changes, this time aimed towards async compute. Basically, this change refactors the way queue families are chosen (and cmdpools are initialized) in order to allow us to make full use of dedicated compute hardware where supported. It also supports a mode (RA_CAP_PARALLEL_COMPUTE) in which it signals to mpv that we have more compute queues than graphics queues, in which case mpv will use compute shaders instead of fragment shaders for most rendering (excluding `finish_pass_fbo`).

In theory, this could help massively on AMD. On nvidia, it's only a very minor slowdown (due to the need to use `CONCURRENT` sharing mode for all resources instead of `EXCLUSIVE`. I still wish there was a way to solve that problem), although that slowdown is somewhat offset by the fact that this lets us trivially use the DMA engine for blit operations (which is slightly faster than using the graphics queue).

Also note that amdgpu currently does not properly export async compute capabilities, so this can only really be tested on AMDGPU-PRO. Does anybody have an AMDGPU-PRO system??