@haasn 2017-09-25
Due to regressions on nvidia implementations caused by the “properly track image lifetime” commit / separate command buffer+semaphores approach, I decided to make that the “cut-off” for revision 1 of vulkan support in mpv, and move that to a separate PR. I'll close this one and re-open it with a cherry-picked stream of changes up to that point (+ later fixes).